# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Desafio QA — Mini Portal",
    "version": "1.0.0",
    "description": "API do Mini Portal (sistema fictício). Os valores de picklist trafegam sempre em português; para exibição, as respostas trazem `<campo>ToLabel` no idioma do usuário (definido no token).\n\nAutenticação: faça login em `POST /api/login`, copie o `token` e informe em **Authorize**.\n\nPermissões por cargo (vêm no JWT em `permissions`): Super Admin (`authorizeds` leitura/escrita, `products` leitura), Proprietário (`contacts` e `employees` leitura/escrita, `products` leitura) e Atendente (`contacts`, `products` e `employees` somente leitura)."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local"
    }
  ],
  "tags": [
    {
      "name": "Sistema"
    },
    {
      "name": "Autenticação"
    },
    {
      "name": "Picklists"
    },
    {
      "name": "Autorizadas"
    },
    {
      "name": "Contatos"
    },
    {
      "name": "Produtos"
    },
    {
      "name": "Funcionários"
    }
  ],
  "paths": {
    "/api/health": {
      "get": {
        "tags": [
          "Sistema"
        ],
        "summary": "Verificação de saúde",
        "responses": {
          "200": {
            "description": "Aplicação no ar",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "ok"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/login": {
      "post": {
        "tags": [
          "Autenticação"
        ],
        "summary": "Login",
        "description": "Valida e-mail e senha e devolve um JWT com cargo, idioma, país, autorizada e permissões.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token gerado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Token"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Validation"
          },
          "401": {
            "description": "Credenciais inválidas",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Error"
                    }
                  ],
                  "example": {
                    "error": "INVALID_CREDENTIALS"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/picklists": {
      "get": {
        "tags": [
          "Picklists"
        ],
        "summary": "Opções dos picklists",
        "description": "Cada opção traz `value` (português, é o que se envia) e `toLabel` (texto no idioma do usuário).",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Picklists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Picklists"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/authorizeds": {
      "get": {
        "tags": [
          "Autorizadas"
        ],
        "summary": "Lista autorizadas",
        "description": "Permissão: `authorizeds` (leitura). Somente Super Admin.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Authorized"
                      }
                    },
                    "total": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      },
      "post": {
        "tags": [
          "Autorizadas"
        ],
        "summary": "Cria autorizada e seu proprietário",
        "description": "Permissão: `authorizeds` (escrita). O proprietário é criado com a senha padrão e herda o país; o idioma é definido pelo país (Brasil: pt; Argentina e Colômbia: es).",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthorizedInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Autorizada criada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Authorized"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Validation"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/DuplicateEmail"
          }
        }
      }
    },
    "/api/contacts": {
      "get": {
        "tags": [
          "Contatos"
        ],
        "summary": "Lista contatos da autorizada",
        "description": "Permissão: `contacts` (leitura). Retorna só os contatos da autorizada do usuário.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Contact"
                      }
                    },
                    "total": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      },
      "post": {
        "tags": [
          "Contatos"
        ],
        "summary": "Cria contato",
        "description": "Permissão: `contacts` (escrita). O telefone segue o formato do país do usuário e é guardado normalizado (`+<código><número>`). O e-mail é único dentro da autorizada.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContactInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Contato criado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Contact"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Validation"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/DuplicateEmail"
          }
        }
      }
    },
    "/api/products": {
      "get": {
        "tags": [
          "Produtos"
        ],
        "summary": "Lista produtos",
        "description": "Permissão: `products` (leitura). Paginado (padrão 10, máximo 50 por página). `name` busca no nome ou no código.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Parte do nome ou do código"
          },
          {
            "name": "category",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "Peças",
                "SKU (White Goods Mercado Nacional)",
                "Modelo Usual (White Goods Mercado Nacional)"
              ]
            },
            "description": "Valor em português"
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "Ativo",
                "Inativo"
              ]
            },
            "description": "Valor em português"
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50,
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Página de produtos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductPage"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Validation"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/api/products/{id}/image": {
      "get": {
        "tags": [
          "Produtos"
        ],
        "summary": "Imagem do produto",
        "description": "Permissão: `products` (leitura). A imagem é guardada no MongoDB.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Imagem PNG",
            "content": {
              "image/png": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "description": "Produto não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Error"
                    }
                  ],
                  "example": {
                    "error": "NOT_FOUND"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/employees": {
      "get": {
        "tags": [
          "Funcionários"
        ],
        "summary": "Lista funcionários da autorizada",
        "description": "Permissão: `employees` (leitura). Retorna só os funcionários da autorizada do usuário.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "role",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "Super Admin",
                "Proprietário",
                "Atendente"
              ]
            },
            "description": "Filtra por cargo (valor em português)"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Employee"
                      }
                    },
                    "total": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Validation"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      },
      "post": {
        "tags": [
          "Funcionários"
        ],
        "summary": "Cria funcionário",
        "description": "Permissão: `employees` (escrita). O funcionário entra na autorizada do usuário, herda país e idioma dela e recebe a senha padrão. O cargo é obrigatório e só aceita Atendente ou Proprietário.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmployeeInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Funcionário criado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Employee"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Validation"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/DuplicateEmail"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Token ausente, inválido ou expirado",
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/Error"
                }
              ],
              "example": {
                "error": "UNAUTHORIZED"
              }
            }
          }
        }
      },
      "Forbidden": {
        "description": "Sem permissão para a ação",
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/Error"
                }
              ],
              "example": {
                "error": "FORBIDDEN"
              }
            }
          }
        }
      },
      "Validation": {
        "description": "Dados inválidos",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
      "DuplicateEmail": {
        "description": "E-mail já cadastrado",
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/ValidationError"
                }
              ],
              "example": {
                "error": "DUPLICATE_EMAIL",
                "fields": {
                  "email": "DUPLICATE_EMAIL"
                }
              }
            }
          }
        }
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "UNAUTHORIZED"
          }
        }
      },
      "ValidationError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "VALIDATION"
          },
          "fields": {
            "type": "object",
            "additionalProperties": {
              "type": "string",
              "enum": [
                "REQUIRED",
                "MIN_LENGTH",
                "INVALID_EMAIL",
                "INVALID_PHONE",
                "INVALID_ROLE",
                "INVALID_COUNTRY",
                "INVALID",
                "DUPLICATE_EMAIL"
              ]
            },
            "example": {
              "email": "INVALID_EMAIL"
            }
          }
        }
      },
      "LoginInput": {
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string",
            "example": "superadmin@example.com"
          },
          "password": {
            "type": "string",
            "example": "Admin@123"
          }
        }
      },
      "Token": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "JWT com sub, email, name, role, roleToLabel, language, country, authorizedId, authorizedName e permissions (ex.: {\"contacts\":\"read/write\"})"
          }
        }
      },
      "PicklistOption": {
        "type": "object",
        "properties": {
          "value": {
            "type": "string",
            "example": "Casa"
          },
          "toLabel": {
            "type": "string",
            "example": "Hogar"
          }
        }
      },
      "Picklists": {
        "type": "object",
        "properties": {
          "category": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PicklistOption"
            }
          },
          "status": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PicklistOption"
            }
          },
          "role": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PicklistOption"
            }
          },
          "assignableRole": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PicklistOption"
            }
          },
          "country": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PicklistOption"
            }
          }
        }
      },
      "Authorized": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "country": {
            "type": "string",
            "enum": [
              "brasil",
              "argentina",
              "colombia"
            ]
          },
          "countryToLabel": {
            "type": "string"
          },
          "language": {
            "type": "string",
            "enum": [
              "pt",
              "es",
              "en"
            ]
          },
          "owner": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string"
              },
              "email": {
                "type": "string"
              }
            }
          },
          "employeesCount": {
            "type": "integer"
          }
        }
      },
      "AuthorizedInput": {
        "type": "object",
        "required": [
          "authorizedName",
          "ownerName",
          "ownerEmail",
          "country"
        ],
        "properties": {
          "authorizedName": {
            "type": "string",
            "minLength": 2
          },
          "ownerName": {
            "type": "string",
            "minLength": 2
          },
          "ownerEmail": {
            "type": "string",
            "format": "email"
          },
          "country": {
            "type": "string",
            "enum": [
              "brasil",
              "argentina",
              "colombia"
            ]
          }
        }
      },
      "Contact": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string",
            "example": "+5511999990000"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ContactInput": {
        "type": "object",
        "required": [
          "name",
          "email",
          "phone"
        ],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "phone": {
            "type": "string",
            "description": "Brasil: 10 ou 11 dígitos; Argentina e Colômbia: 10 dígitos (com ou sem o código do país)",
            "example": "(11) 99999-0000"
          }
        }
      },
      "Product": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "code": {
            "type": "string",
            "example": "K51720601"
          },
          "name": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "category": {
            "type": "string",
            "enum": [
              "Peças",
              "SKU (White Goods Mercado Nacional)",
              "Modelo Usual (White Goods Mercado Nacional)"
            ]
          },
          "categoryToLabel": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "Ativo",
              "Inativo"
            ]
          },
          "statusToLabel": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "imageUrl": {
            "type": "string",
            "example": "/api/products/{id}/image"
          }
        }
      },
      "ProductPage": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Product"
            }
          },
          "total": {
            "type": "integer"
          },
          "page": {
            "type": "integer"
          },
          "pageSize": {
            "type": "integer"
          }
        }
      },
      "Employee": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "Super Admin",
              "Proprietário",
              "Atendente"
            ]
          },
          "roleToLabel": {
            "type": "string"
          },
          "language": {
            "type": "string",
            "enum": [
              "pt",
              "es",
              "en"
            ]
          },
          "country": {
            "type": "string",
            "enum": [
              "brasil",
              "argentina",
              "colombia"
            ]
          }
        }
      },
      "EmployeeInput": {
        "type": "object",
        "required": [
          "name",
          "email",
          "role"
        ],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "role": {
            "type": "string",
            "enum": [
              "Atendente",
              "Proprietário"
            ]
          }
        }
      }
    }
  }
}